Split zenodo_metadata git tag tests into unit + integration
The verify_git_tag_checked_out tests that build a throwaway git repo and
shell out to real git are fast, but conceptually belong under
tests/integration for developer legibility. Move them (and the git_repo
fixture) verbatim into a new tests/integration/deploy/zenodo_metadata_test.py.

Replace them in tests/unit/deploy/zenodo_metadata_test.py with a single
parametrized test that mocks run_git and covers the three outcomes of the
tag-vs-HEAD comparison logic: matching commits, mismatched commits, and an
unresolvable tag.

Also trim a stale sentence from the run_git docstring.
